refactor: move files into app
