+++ app/api/auth/[...nextauth]/route.ts
@@ -1,0 +1,6 @@
+import { authOptions } from "@/app/lib/auth";
+import NextAuth from "next-auth";
+
+const handler = NextAuth(authOptions);
+
+export { handler as GET, handler as POST };
