+++ app/components/averageInPrice/StyledButton.tsx
