+++ app/types/stockTypes.ts
