+++ app/utils/formatNumber.ts
